setting wrappers window bounds
