+++ docking-frames-common/pom.xml
@@ -22,4 +22,8 @@
 			<version>${project.version}</version>
 		</dependency>
 	</dependencies>
+    <properties>
+        <maven.compiler.source>1.8</maven.compiler.source>
+        <maven.compiler.target>1.8</maven.compiler.target>
+    </properties>
 </project>
